fix: comment 삭제 관련 예외처리를 추가한다

+++ backend/src/main/kotlin/com/dclass/backend/application/CommentService.kt
@@ -53,6 +53,7 @@ class CommentService(
     fun delete(userId: Long, request: DeleteCommentRequest) {
         val comment = commentRepository.findByIdAndUserId(request.commentId, userId)
             ?: throw CommentException(CommentExceptionType.NOT_FOUND_COMMENT)
+        if(comment.isDeleted()) throw CommentException(CommentExceptionType.DELETED_COMMENT)
         commentRepository.delete(comment)
 
         val post = postRepository.findByIdOrThrow(comment.postId)
